Added CommandTimers feature and config.yml
CommandTimers is a plugin-based solution to having too many ticking
command blocks on a server. Rather than have a bunch of repeating
blocks that execute every tick, you can use this to build efficient
clocked systems

This system works by placing armor stands with the appropriate tags on
top of impulse or repeating command blocks. The plugin automatically
runs the commands according to the parameters set by tags on the armor
stand.

Tags:
- timer (required)
- period=### (time in ticks, required)
- range=### (only active when player is within ### blocks, optional)
- repeat
  - This timer is for a repeating command block
  - Repeating block underneath will be turned on and off as players move
    within range
  - If specified, range must also be specified

Signed-off-by: Byron Marohn <combustible@live.com>